# ~2x speed up in serialization of `java.time.*` instances by default encoders