Bellow are results of benchmarks which writes an array of 128 `java.time.*` instances as a JSON array of strings:

##Before
```
[info] Benchmark                            (size)   Mode  Cnt      Score      Error  Units
[info] ArrayOfLocalDateTimesWriting.circe      128  thrpt    5  17828.304 ± 2224.497  ops/s
[info] ArrayOfLocalTimesWriting.circe          128  thrpt    5  28257.429 ± 2637.814  ops/s
[info] ArrayOfOffsetDateTimesWriting.circe     128  thrpt    5  14812.471 ±  745.845  ops/s
[info] ArrayOfOffsetTimesWriting.circe         128  thrpt    5  21762.077 ± 2176.789  ops/s
[info] ArrayOfZonedDateTimesWriting.circe      128  thrpt    5  12016.890 ± 1583.515  ops/s
```
##After
```
[info] Benchmark                            (size)   Mode  Cnt      Score      Error  Units
[info] ArrayOfLocalDateTimesWriting.circe      128  thrpt    5  29355.954 ± 2920.523  ops/s
[info] ArrayOfLocalTimesWriting.circe          128  thrpt    5  58263.162 ± 8294.110  ops/s
[info] ArrayOfOffsetDateTimesWriting.circe     128  thrpt    5  25150.080 ± 2462.175  ops/s
[info] ArrayOfOffsetTimesWriting.circe         128  thrpt    5  41273.528 ± 7187.940  ops/s
[info] ArrayOfZonedDateTimesWriting.circe      128  thrpt    5  20173.697 ± 2366.075  ops/s
```