@plokhotnyuk 2019-09-19
Benchmarks above shows improvement for serialization only.

Parsing speed doesn't changed, but using of formatters was removed with hope that default `parse` functions will be revised in future versions of Java for better performance...